feat: custom border
Creating the 'custom_border' file on utils folder, to organize the current custom borders on the app.
